@@ -8,21 +8,24 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2.5",
-        "fideloper/proxy": "4.4.1",
-        "fruitcake/laravel-cors": "2.0.3",
-        "guzzlehttp/guzzle": "7.2.0",
-        "laravel/framework": "7.30.2",
-        "laravel/tinker": "2.5.0",
-        "laravel/ui": "^2.1",
+        "php": "^7.3|^8.0",
+        "doctrine/dbal": "^3.0",
+        "fideloper/proxy": "^4.4",
+        "fruitcake/laravel-cors": "^2.0",
+        "guzzlehttp/guzzle": "^7.0.1",
+        "laravel/framework": "^8.12",
+        "laravel/sanctum": "^2.8",
+        "laravel/tinker": "^2.5",
+        "laravel/ui": "^3.0",
         "lorisleiva/laravel-deployer": "0.3.5",
         "sentry/sentry-laravel": "^2.0"
     },
     "require-dev": {
-        "facade/ignition": "2.5.8",
-        "fzaninotto/faker": "1.9.2",
-        "mockery/mockery": "1.4.2",
-        "nunomaduro/collision": "5.2.0",
+        "facade/ignition": "^2.5",
+        "fakerphp/faker": "^1.9.1",
+        "laravel/sail": "^1.0.1",
+        "mockery/mockery": "^1.4.2",
+        "nunomaduro/collision": "^5.0",
         "pestphp/pest": "^1.0",
         "pestphp/pest-plugin-laravel": "^1.0",
         "phpunit/phpunit": "9.4.4"
@@ -42,12 +45,10 @@
             "app/Http/helpers.php"
         ],
         "psr-4": {
-            "App\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
