@@ -2,36 +2,36 @@
 
 namespace Tests\Feature;
 
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Illuminate\Foundation\Testing\WithFaker;
 use Tests\TestCase;
-use App\User;
-use App\News;
+use App\Models\User;
+use App\Models\News;
 
 class NewsTest extends TestCase
 {
+    private $user;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+        $this->user = User::factory()->create();
+    }
+
     public function testNewsAreCreatedCorrectly()
     {
-        $user = factory(User::class)->create();
-        $token = $user->generateToken();
-        $headers = ['Authorization' => "Bearer $token"];
         $payload = [
             'title' => 'Lorem',
             'text' => 'Ipsum',
             'date' => '2020-08-04',
         ];
 
-        $this->json('POST', '/api/news', $payload, $headers)
+        $this->actingAs($this->user)->postJson('/api/news', $payload)
             ->assertStatus(201)
             ->assertJson(['title' => 'Lorem', 'text' => 'Ipsum']);
     }
 
     public function testNewsAreUpdatedCorrectly()
     {
-        $user = factory(User::class)->create();
-        $token = $user->generateToken();
-        $headers = ['Authorization' => "Bearer $token"];
-        $news = factory(News::class)->create([
+        $news = News::factory()->create([
             'title' => 'First News',
             'text' => 'First Body',
             'date' => '2020-08-04',
@@ -43,7 +43,7 @@ class NewsTest extends TestCase
             'date' => '2020-08-04',
         ];
 
-        $response = $this->json('PUT', '/api/news/' . $news->ID, $payload, $headers)
+        $this->actingAs($this->user)->putJson('/api/news/' . $news->ID, $payload)
             ->assertStatus(200)
             ->assertJson([
                 'ID' => $news->ID,
@@ -54,36 +54,29 @@ class NewsTest extends TestCase
 
     public function testNewsAreDeletedCorrectly()
     {
-        $user = factory(User::class)->create();
-        $token = $user->generateToken();
-        $headers = ['Authorization' => "Bearer $token"];
-        $news = factory(News::class)->create([
+        $news = News::factory()->create([
             'title' => 'First News',
             'text' => 'First Body',
             'date' => '2020-08-04',
         ]);
 
-        $this->json('DELETE', '/api/news/' . $news->ID, [], $headers)
+        $this->actingAs($this->user)->deleteJson('/api/news/' . $news->ID, [])
             ->assertStatus(204);
     }
 
     public function testNewsAreListedCorrectly()
     {
-        factory(News::class)->create([
+        News::factory()->create([
             'title' => 'First News',
             'text' => 'First Body'
         ]);
 
-        factory(News::class)->create([
+        News::factory()->create([
             'title' => 'Second News',
             'text' => 'Second Body'
         ]);
 
-        $user = factory(User::class)->create();
-        $token = $user->generateToken();
-        $headers = ['Authorization' => "Bearer $token"];
-
-        $response = $this->json('GET', '/api/news', [], $headers)
+        $this->actingAs($this->user)->getJson('/api/news', [])
             ->assertStatus(200)
             ->assertJsonStructure([
                 '*' => ['ID', 'title', 'text', 'date', 'expirationDate'],
