@@ -2,11 +2,11 @@
 
 namespace Tests\Feature;
 
-use App\Navigation;
+use App\Models\Page;
+use App\Models\Navigation;
 use App\Router\Router;
 use Illuminate\Foundation\Testing\RefreshDatabase;
 use Tests\TestCase;
-use Cache;
 
 class RouterTest extends TestCase
 {
@@ -29,11 +29,11 @@ class RouterTest extends TestCase
     /** @test */
     public function it_should_find_a_page_by_url()
     {
-        $pageOne = factory(\App\Page::class)->create();
-        $pageTwo = factory(\App\Page::class)->create();
+        $pageOne = Page::factory()->create();
+        $pageTwo = Page::factory()->create();
 
-        $navOne = factory(\App\Navigation::class)->create(['refID' => $pageOne->id]);
-        $navTwo = factory(\App\Navigation::class)->create(['refID' => $pageTwo->id]);
+        $navOne = Navigation::factory()->create(['refID' => $pageOne->id]);
+        $navTwo = Navigation::factory()->create(['refID' => $pageTwo->id]);
 
         $page = $this->router->findByUrl($navOne->slug);
         $this->assertNotNull($page);
@@ -43,11 +43,11 @@ class RouterTest extends TestCase
     /** @test */
     public function it_should_not_find_a_non_existing_page()
     {
-        $pageOne = factory(\App\Page::class)->create();
-        $pageTwo = factory(\App\Page::class)->create();
+        $pageOne = Page::factory()->create();
+        $pageTwo = Page::factory()->create();
 
-        $navOne = factory(\App\Navigation::class)->create(['refID' => $pageOne->id, 'linkname' => 'nav-one']);
-        $navTwo = factory(\App\Navigation::class)->create(['refID' => $pageTwo->id, 'linkname' => 'nav-two']);
+        $navOne = Navigation::factory()->create(['refID' => $pageOne->id, 'linkname' => 'nav-one']);
+        $navTwo = Navigation::factory()->create(['refID' => $pageTwo->id, 'linkname' => 'nav-two']);
 
         $page = $this->router->findByUrl('foo-bar');
         $this->assertNull($page);
