@@ -3,68 +3,52 @@
 namespace App\Http\Controllers\Auth;
 
 use App\Http\Controllers\Controller;
-use App\Providers\RouteServiceProvider;
-use Illuminate\Foundation\Auth\AuthenticatesUsers;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
-use Carbon\Carbon;
 
 class LoginController extends Controller
 {
-    /*
-    |--------------------------------------------------------------------------
-    | Login Controller
-    |--------------------------------------------------------------------------
-    |
-    | This controller handles authenticating users for the application and
-    | redirecting them to your home screen. The controller uses a trait
-    | to conveniently provide its functionality to your applications.
-    |
-    */
-
-    use AuthenticatesUsers;
-
     /**
-     * Where to redirect users after login.
+     * Handle an authentication attempt.
      *
-     * @var string
+     * @param  \Illuminate\Http\Request $request
+     * @return \Illuminate\Http\Response
      */
-    protected $redirectTo = RouteServiceProvider::HOME;
-
-    /**
-     * Create a new controller instance.
-     *
-     * @return void
-     */
-    public function __construct()
+    public function authenticate(Request $request)
     {
-        $this->middleware('guest')->except('logout');
-    }
+        $this->validate($request, [
+            'email' => 'required|email',
+            'password' => 'required',
+        ]);
 
-    public function login(Request $request)
-    {
-        $this->validateLogin($request);
+        $credentials = $request->only('email', 'password');
 
-        if ($this->attemptLogin($request)) {
-            $user = $this->guard()->user();
-            $user->generateToken();
+        if (Auth::attempt($credentials)) {
+            $request->session()->regenerate();
 
-            return response()->json([
-                'user' => $user->toArray(),
-            ]);
+            return response()->json(Auth::user());
         }
 
-        return $this->sendFailedLoginResponse($request);
+        return response()->json([
+            'errors' => [
+                'email' => ['auth.failed']
+            ]
+        ], 401);
     }
 
+    /**
+     * Log the user out of the application.
+     *
+     * @param  \Illuminate\Http\Request $request
+     * @return \Illuminate\Http\Response
+     */
     public function logout(Request $request)
     {
-        $user = Auth::guard('api')->user();
+        Auth::logout();
 
-        if ($user) {
-            $user->api_token = null;
-            $user->save();
-        }
+        $request->session()->invalidate();
+
+        $request->session()->regenerateToken();
 
         return response()->json(['data' => 'User logged out.'], 200);
     }
