@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-use App\News;
+use App\Models\News;
 use Illuminate\Http\Request;
 
 class NewsController extends Controller
