@@ -2,7 +2,7 @@
 
 namespace App\Http\Middleware;
 
-use App\Role;
+use App\Models\Role;
 use Closure;
 use Illuminate\Support\Facades\Auth;
 
@@ -17,8 +17,8 @@ class IsAdmin
      */
     public function handle($request, Closure $next)
     {
-        if(!Auth::user()->hasRole(Role::ADMIN)){
-            return response()->json(['error'=>'Unauthorised'], 401);
+        if (!Auth::user()->hasRole(Role::ADMIN)) {
+            return response()->json(['error' => 'Unauthorised'], 401);
         }
 
         return $next($request);
