@@ -1,11 +1,14 @@
 <?php
 
-namespace App;
+namespace App\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 
 class News extends Model
 {
+    use HasFactory;
+
     protected $table = 'tbl_news';
 
     protected $primaryKey = 'ID';
