@@ -1,9 +1,8 @@
-import Config from "../config";
 import News from "../models/news";
 import BaseService from "./base.service";
 
 class NewsService extends BaseService<News> {
-    baseUrl = `${Config.host}/api/news`;
+    baseUrl = "/api/news";
 }
 
 export default new NewsService();
