@@ -1,9 +1,8 @@
-import Config from "../config";
 import Page from "../models/page";
 import BaseService from "./base.service";
 
 class PageService extends BaseService<Page> {
-    baseUrl = `${Config.host}/api/pages`;
+    baseUrl = "/api/pages";
 }
 
 export default new PageService();
