@@ -1,13 +1,12 @@
 import User from "../models/user";
-import Config from "../config";
 import BaseService from "./base.service";
 
 class UserService extends BaseService<User> {
-    baseUrl = `${Config.host}/api/users`;
+    baseUrl = "/api/users";
 
     revoke(user: User): Promise<void> {
         return this.apiClient
-            .post(`${this.baseUrl}/${user.id}/revoke`, {}, this.defaultOptions)
+            .post(`${this.baseUrl}/${user.id}/revoke`, {})
             .then(response => response.data);
     }
 }
