@@ -2,7 +2,7 @@
     <div class="container mt-3 p-3">
         <div class="row">
             <div class="col">
-                <h1>Willkommen, {{ user.name }}</h1>
+                <h1>Willkommen, {{ user ? user.name : "" }}</h1>
                 <p>
                     Wählen Sie einen Menüpunkt aus um den jeweiligen Bereich zu
                     bearbeiten
